+++ R/forest.netcomb.R
@@ -167,7 +167,7 @@ forest.netcomb <- function(x,
   ##
   pooled <- meta:::setchar(pooled, c("fixed", "random"))
   ##
-  meta:::chknumeric(digits, min = 0, single = TRUE)
+  meta:::chknumeric(digits, min = 0, length = 1)
   ##
   chklogical(baseline.reference)
   chklogical(drop.reference.group)