+++ R/forest.netmeta.R
@@ -214,7 +214,7 @@ forest.netmeta <- function(x,
   ##
   pooled <- meta:::setchar(pooled, c("fixed", "random"))
   ##
-  meta:::chknumeric(digits, min = 0, single = TRUE)
+  meta:::chknumeric(digits, min = 0, length = 1)
   ##
   chklogical(baseline.reference)
   ##