+++ R/funnel.netmeta.R
@@ -239,11 +239,11 @@ funnel.netmeta <- function(x,
   chkchar(text.mm)
   ##
   chkchar(sep.trts)
-  chknumeric(nchar.trts, min = 1, single = TRUE)
+  chknumeric(nchar.trts, min = 1, length = 1)
   ##
   chklogical(backtransf)
   ##
-  chknumeric(digits.pval, min = 1, single = TRUE)
+  chknumeric(digits.pval, min = 1, length = 1)
   
   
   ##