+++ R/netimpact.R
@@ -64,8 +64,8 @@ netimpact <- function(x, seTE.ignore = 1e4, event.ignore = 0.01, verbose = FALSE
   
   meta:::chkclass(x, "netmeta")
   ##
-  meta:::chknumeric(seTE.ignore, min = 0, zero = TRUE, single = TRUE)
-  meta:::chknumeric(event.ignore, min = 0, zero = TRUE, single = TRUE)
+  meta:::chknumeric(seTE.ignore, min = 0, zero = TRUE, length = 1)
+  meta:::chknumeric(event.ignore, min = 0, zero = TRUE, length = 1)
   
   
   studlab <- x$studlab